backend/requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.1
pydantic-settings==2.8.1
chromadb==0.6.3
sentence-transformers==3.4.1
ollama==0.4.7
python-dotenv==1.0.1
